+++ src/main/java/io/github/maritims/NpmMojo.java
@@ -1,6 +1,6 @@
-package com.github.maritims;
+package io.github.maritims;
 
-import com.github.maritims.node.NodeConfiguration;
+import io.github.maritims.node.NodeConfiguration;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
@@ -9,7 +9,6 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.project.MavenProject;
 
-import java.io.IOException;
 import java.nio.file.Paths;
 
 @Mojo(name = "npm", defaultPhase = LifecyclePhase.INSTALL)
@@ -44,11 +43,7 @@ public class NpmMojo extends AbstractMojo {
                 Paths.get(project.getBasedir().getAbsolutePath(), "src", "main", sourceCodeDirectoryName).toString()
         );
 
-        try {
-            if(install) npm.install();
-            npm.run(script);
-        } catch (IOException | InterruptedException e) {
-            throw new RuntimeException(e);
-        }
+        if(install) npm.install();
+        npm.run(script);
     }
 }